+++ include/pacmod3/pacmod3_ros_msg_handler.h
@@ -35,7 +35,7 @@ private:
 class Pacmod3TxRosMsgHandler
 {
 public:
-  void fillAndPublish(const int64_t& can_id,
+  void fillAndPublish(const uint32_t& can_id,
                       const std::string& frame_id,
                       const ros::Publisher& pub,
                       const std::shared_ptr<Pacmod3TxMsg>& parser_class);
