+++ src/pacmod3_ros_msg_handler.cpp
@@ -32,7 +32,7 @@ void LockedData::setData(std::vector<unsigned char> new_data)
 }
 
 void Pacmod3TxRosMsgHandler::fillAndPublish(
-    const int64_t& can_id,
+    const uint32_t& can_id,
     const std::string& frame_id,
     const ros::Publisher& pub,
     const std::shared_ptr<Pacmod3TxMsg>& parser_class)
